+++ ChangeLog.md
@@ -5,6 +5,7 @@
   - Add static VersionInfo GetVersion(const QString& szVersion);
   - Add static bool VersionValid(const QString& szVersion);
   - Add static int VersionCompare(const QString &ver1, const QString &ver2);
+  - Add password generator
 - Add set applicaiont font
 - Add audio and video settings widget
 - Add CIconUtils